Old databases and indexes are obsolete.
