+++ src/ContentRepository.InMemory/InitialTestData.cs
@@ -1,11 +1,14 @@
 ﻿// Generated by a tool. 2020-01-14
 // sensenet version: 7.7.5.0
+
+using System;
 using SenseNet.ContentRepository.Storage.DataModel;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 
 namespace SenseNet.ContentRepository.InMemory
 {
+    [Obsolete("Use InMemoryTestDatabase.", true)]
     public class InitialTestData : IRepositoryDataFile
     {
         private InitialTestData()
