+++ src/ContentRepository/DefaultDatabase.cs
@@ -1,11 +1,14 @@
 ﻿// Generated by a tool. 2020-01-16
 // sensenet version: 7.7.5.1
+
+using System;
 using SenseNet.ContentRepository.Storage.DataModel;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 
 namespace SenseNet.ContentRepository.InMemory
 {
+    [Obsolete("Use the SenseNetServicesData.", true)]
     public class DefaultDatabase : IRepositoryDataFile
     {
         private DefaultDatabase()
