+++ src/Tests/SenseNet.ContentRepository.Tests/SenseNet.ContentRepository.Tests.csproj
@@ -153,7 +153,6 @@
     <Compile Include="LoggerTests.cs" />
     <Compile Include="MaintenanceTests.cs" />
     <Compile Include="NodeTests.cs" />
-    <Compile Include="InMemoryDatabaseGenerator.cs" />
     <Compile Include="UnknownContentHandlerTests.cs" />
     <Compile Include="SharedLockTests.cs" />
     <Compile Include="SharingTests.cs" />
